The MicrobiomeExperiment class

The MicrobiomeExperiment class is a SummarizedExperiment-like class for microbiome data inheriting from TreeSummarizedExperiment and SingleCellExperiment.

MicrobiomeExperiment(..., referenceSeq = NULL)

Arguments

...
Arguments passed to TreeSummarizedExperiment

referenceSeq
a DNAStringSet object or some object coercible to a DNAStringSet object. See DNAStringSet for more details.

Details

MicrobiomeExperiment adds to the above mentioned class a slot for storing a reference sequence as a DNAStringSet.

Examples

data(taxa)
sampleNames <- letters[1:4]
pd <- DataFrame(a=letters[1:4], b=1:4)
numcounts <- nrow(taxa) * 4
counts <- matrix(sample(1:1000, numcounts, replace=TRUE),
                 nr = nrow(taxa), nc = 4)
refSeq <- DNAStringSetList(one = DNAStringSet(c("A","A","A","A","A")),
                           two = DNAStringSet(c("A","A","A","A","A")))
me <- MicrobiomeExperiment(assays = SimpleList(counts = counts),
                           rowData = taxa,
                           colData = pd,
                           referenceSeq = refSeq)
me

#> class: MicrobiomeExperiment 
#> dim: 5 4 
#> metadata(0):
#> assays(1): counts
#> rownames: NULL
#> rowData names(9): Keys OTU ... Genus Species
#> colnames: NULL
#> colData names(2): a b
#> reducedDimNames(0):
#> altExpNames(0):
#> rowLinks: NULL
#> rowTree: NULL
#> colLinks: NULL
#> colTree: NULL
#> referenceSeq: a DNAStringSetList (2 x 5 sequences each)